pyproject.toml
[project]
name = "quirkink-razrabot"
package-mode = false
version = "0.0.1"
description = "A test task for RAZRABOT"
authors = [
    {name = "Mikhail Shevelkov"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "aiohttp (>=3.11.14,<4.0.0)",
    "quirkink-bot @ git+https://{{ TOKEN }}:x-oauth-basic@github.com/Quirkink/quirkink-bot@1.0.6",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "openai (>=1.68.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
